requirements.txt
absl-py==2.1.0
accelerate==0.28.0
aiosignal==1.3.1
altair==5.2.0
AMFM_decompy==1.0.11
amqp==5.2.0
annotated-types==0.6.0
apprise==1.7.4
asn1crypto==1.5.1
astral==3.2
astunparse==1.6.3
Automat==22.10.0
bcrypt==4.1.2
beautifulsoup4==4.12.2
billiard==4.2.0
bleach==6.1.0
blinker==1.7.0
boto3==1.28.72
botocore==1.31.72
cbor2==5.6.2
celery==5.3.4
certifi==2024.2.2
chardet==3.0.4
constantly==23.10.4
contourpy==1.2.0
crepe==0.0.14
crispy-bootstrap4==2023.1
cssselect==1.2.0
cycler==0.12.1
Cython==3.0.9
decorator==4.4.2
deepface==0.0.74
defusedxml==0.7.1
Django==5.0.3
django-address==0.2.8
django-allauth==0.61.1
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-jsignature==0.11
django-mfa2==2.8.0
django-pwa-webpush==1.1.3
django-recaptcha==4.0.0
django-storages==1.14.2
django-summernote==0.8.20.0
django-webauth==0.1.2
dnspython==2.6.1
dsplib==0.9.0
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
facexlib==0.3.0
filelock==3.13.1
filterpy==1.4.5
Flask==3.0.2
flatbuffers==24.3.7
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
future==1.0.0
gast==0.5.4
gdown==5.1.0
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.18.0
google-api-python-client==2.123.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googleapis-common-protos==1.63.0
googletrans==3.0.0
grpcio==1.62.1
h11==0.9.0
h2==3.2.0
h3==3.7.7
h5py==3.10.0
hmmlearn==0.3.2
hpack==3.0.0
hstspreload==2024.3.1
http_ece==1.2.0
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
huggingface-hub==0.20.3
hyperframe==5.2.0
hyperlink==21.0.0
hypothesis==6.99.13
idna==3.6
ijson==3.2.3
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib-metadata==6.11.0
imutils==0.5.4
incremental==22.10.0
iniconfig==2.0.0
itsdangerous==2.1.2
jellyfish==1.0.3
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
JPype1==1.3.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==3.1.1
kiwisolver==1.4.5
kombu==5.3.5
langdetect==1.0.9
lazy_loader==0.3
libclang==18.1.1
llvmlite==0.42.0
lmdb==1.4.1
loguru==0.5.3
lxml==5.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
ml-dtypes==0.3.2
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
mtcnn==0.1.1
multidict==6.0.5
namex==0.0.7
networkx==3.2.1
numba==0.59.1
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.9.6.50
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
openai-clip==1.0.1
opt-einsum==3.3.0
optree==0.10.0
packaging==24.0
pandas==2.2.1
pdf2image==1.16.0
pdf417==0.8.1
penn==0.0.13
Pillow==9.5.0
pitch-detectors==0.5.2
platformdirs==4.2.0
pluggy==1.4.0
pooch==1.8.1
praat-parselmouth==0.4.3
priority==1.3.0
proglog==0.1.10
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
psycopg2==2.9.9
py==1.11.0
pyaml==23.12.0
pyarrow==15.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycares==4.4.0
pycparser==2.21
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.16.1
pyheif==0.5.1
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==24.1.0
pyotp==2.9.0
pyparsing==3.1.2
pyquery==2.0.0
pyreaper==0.0.10
PySocks==1.7.1
pysptk==0.2.2
pytesseract==0.3.10
pytest==6.2.5
pytest-datadir==1.5.0
pytest-regressions==2.2.0
python-dateutil==2.8.2
python-jose==3.3.0
python3-openid==3.2.0
pytz==2024.1
pyworld==0.3.4
PyYAML==6.0.1
realesrgan==0.3.0
redis==4.6.0
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
resampy==0.2.2
retina-face==0.0.14
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
s3transfer==0.7.0
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.13.0rc1
sentry-sdk==1.40.5
sewar==0.4.6
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
SpeechRecognition==3.10.0
sqlparse==0.4.4
srt==3.5.3
streamlit==1.27.0
stripe==7.0.0
sympy==1.12
tb-nightly==2.17.0a20240323
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorrt==8.6.1.post1
tensorrt-bindings==8.6.1
termcolor==2.4.0
tf_keras==2.16.0
threadpoolctl==3.4.0
tifffile==2024.2.12
timezonefinder==6.4.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.2.1
torch-yin==0.1.3
torchaudio==2.2.1
torchcrepe==0.0.22
torchutil==0.0.11
torchvision==0.17.1
tornado==6.4
tqdm==4.66.2
triton==2.2.0
twilio==9.0.2
Twisted==24.3.0
txaio==23.1.1
typer==0.4.0
tzdata==2024.1
tzlocal==5.2
ua-parser==0.18.0
uritemplate==4.1.1
uritools==4.0.2
urlextract==1.9.0
user-agents==2.2.0
validators==0.24.0
vine==5.1.0
vosk==0.3.45
watchdog==4.0.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.1
whatimage==0.0.3
wrapt==1.16.0
yapecs==0.0.7
yapf==0.40.2
yarl==1.9.4
zipp==3.18.1
zope.interface==6.2
zxingmod==0.19
autocorrect
paramiko
ccxt
numpy==1.26.0
urllib3
webauthn
docbarcodes
